Added relation, README.md, logs and tests. Sorry for long time waiting. I'm from Windows 7 :(

+++ logs_and_tests/src/main/java/ua/com/alevel/department/DepartmentService.java
@@ -28,7 +28,12 @@ public class DepartmentService {
         LOGGER_WARN.warn("End update department");
     }
     public void delete(String id) {
-        departmentDao.delete(id);
+        if (id != null) {
+            departmentDao.delete(id);
+        }
+        else {
+            LOGGER_ERROR.error("This id does not exist");
+        }
     }
 
     public Department findById(String id) {
